requirements.txt
matplotlib==3.10.1
numpy
pandas
Pillow==11.1.0
py-feat==0.7.0
streamlit==1.43.0
opencv-python-headless